Consolidate type families and condition negation (review findings 3-4)
TypeFamilies is the single home for stack-family classification — the
knowledge previously lived in four shapes across the importer's slot
merging (enum + switch), the structuring pass's float detection
(nullable bool), and three printer spellings (unsigned casts,
truthiness, float checks). Conditions is the single home for the
type-aware negation duals; the structuring pass delegates, and the
printer's LogicalNot-over-comparison fold now routes through the
shared inverse WITH the float unordered-flag flip it previously
lacked — flat (unstructured) brfalse over a float ordering comparison
now reaches the NaN-correct !(a < b) spelling too.

This is also where IsValueType knowledge lands when resolution
provides it, closing the receiver-purity and truthiness-unknown gaps
in one place.

Behavior identical except the flat-float-fold correctness fix: all
suites green both configs, parity steady at 40.60%.
